requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic[email]
streamlit
groq
